Added Publisher Controller CRUD

+++ Libreria.BLL/DTOs/PublisherDTO.cs
@@ -10,7 +10,7 @@ namespace BibliotecaBLL.DTOs
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public byte YearOfPublication { get; set; }
+        public short YearOfPublication { get; set; }
         public string WebSite { get; set; }
         public string Email { get; set; }
     }
